fix multishot, add piercing effect

+++ src/main/java/net/smileycorp/raids/common/item/ItemCrossbow.java
@@ -12,14 +12,12 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.entity.projectile.EntityArrow;
 import net.minecraft.init.Items;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemArrow;
-import net.minecraft.item.ItemFirework;
-import net.minecraft.item.ItemStack;
+import net.minecraft.item.*;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.nbt.NBTTagList;
 import net.minecraft.stats.StatList;
 import net.minecraft.util.*;
+import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.util.text.Style;
 import net.minecraft.util.text.TextComponentString;
@@ -28,6 +26,7 @@ import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.World;
 import net.minecraftforge.event.ForgeEventFactory;
 import net.smileycorp.raids.common.Constants;
+import net.smileycorp.raids.common.MathUtils;
 import net.smileycorp.raids.common.RaidsContent;
 import net.smileycorp.raids.common.RaidsSoundEvents;
 import net.smileycorp.raids.common.entities.ICrossbowArrow;
@@ -35,6 +34,7 @@ import net.smileycorp.raids.common.entities.ICrossbowAttackMob;
 import net.smileycorp.raids.common.entities.IFireworksProjectile;
 
 import javax.vecmath.Vector3f;
+import javax.vecmath.Vector4f;
 import java.util.List;
 import java.util.Random;
 
@@ -61,7 +61,7 @@ public class ItemCrossbow extends Item {
 	public ActionResult<ItemStack> onItemRightClick(World world, EntityPlayer player, EnumHand hand) {
 		ItemStack stack = player.getHeldItem(hand);
 		if (isCharged(stack)) {
-			performShooting(world, player, hand, stack, containsChargedProjectile(stack, Items.FIREWORKS) ? 1.6F : 3.15F, 1.0F);
+			performShooting(world, player, stack, containsChargedProjectile(stack, Items.FIREWORKS) ? 1.6F : 3.15F, 1.0F);
 			setCharged(stack, false);
 			return ActionResult.newResult(EnumActionResult.SUCCESS, stack);
 		} else if (!getProjectile(player).isEmpty()) {
@@ -92,40 +92,26 @@ public class ItemCrossbow extends Item {
 		boolean creative = entity instanceof EntityPlayer && ((EntityPlayer)entity).capabilities.isCreativeMode;
 		ItemStack itemstack = getProjectile(entity);
 		ItemStack itemstack1 = itemstack.copy();
-
 		for(int k = 0; k < j; ++k) {
-			if (k > 0) {
-				itemstack = itemstack1.copy();
-			}
-
+			if (k > 0) itemstack = itemstack1.copy();
 			if (itemstack.isEmpty() && creative) {
 				itemstack = new ItemStack(Items.ARROW);
 				itemstack1 = itemstack.copy();
 			}
-
-			if (!loadProjectile(entity, stack, itemstack, k > 0, creative)) {
-				return false;
-			}
+			if (!loadProjectile(entity, stack, itemstack, k > 0, creative)) return false;
 		}
-
 		return true;
 	}
 
 	private static boolean loadProjectile(EntityLivingBase entity, ItemStack stack, ItemStack ammo, boolean p_40866_, boolean creative) {
-		if (ammo.isEmpty()) {
-			return false;
-		} else {
+		if (ammo.isEmpty()) return false;
+		else {
 			boolean dontConsume = creative && ammo.getItem() instanceof ItemArrow;
 			ItemStack itemstack;
 			if (!dontConsume && !creative && !p_40866_) {
 				itemstack = ammo.splitStack(1);
-				if (ammo.isEmpty() && entity instanceof EntityPlayer) {
-					((EntityPlayer)entity).inventory.deleteStack(stack);
-				}
-			} else {
-				itemstack = ammo.copy();
-			}
-
+				if (ammo.isEmpty() && entity instanceof EntityPlayer) ((EntityPlayer)entity).inventory.deleteStack(stack);
+			} else itemstack = ammo.copy();
 			addChargedProjectile(stack, itemstack);
 			return true;
 		}
@@ -148,11 +134,8 @@ public class ItemCrossbow extends Item {
 		NBTTagCompound nbt = stack.getTagCompound();
 		if (nbt == null) nbt = new NBTTagCompound();
 		NBTTagList taglist;
-		if (nbt.hasKey("ChargedProjectiles", 9)) {
-			taglist = nbt.getTagList("ChargedProjectiles", 10);
-		} else {
-			taglist = new NBTTagList();
-		}
+		if (nbt.hasKey("ChargedProjectiles", 9)) taglist = nbt.getTagList("ChargedProjectiles", 10);
+		else taglist = new NBTTagList();
 		NBTTagCompound stacknbt = new NBTTagCompound();
 		projectile.writeToNBT(stacknbt);
 		taglist.appendTag(stacknbt);
@@ -179,71 +162,57 @@ public class ItemCrossbow extends Item {
 		return getChargedProjectiles(stack).stream().anyMatch(proj -> proj.getItem() == item);
 	}
 
-	private static void shootProjectile(World world, EntityLivingBase entity, EnumHand hand, ItemStack stack, ItemStack ammo, float p_40900_, boolean p_40901_, float p_40902_, float p_40903_, float p_40904_) {
-		if (!world.isRemote) {
-			boolean isFirework = ammo.getItem() == Items.FIREWORKS;
-			Entity projectile;
-			if (isFirework) {
-				projectile = new EntityFireworkRocket(world, entity.posX, entity.posY + entity.getEyeHeight() - 0.15F, entity.posZ, ammo);
-				((IFireworksProjectile)projectile).setOwner(entity);
-				((IFireworksProjectile)projectile).setShotAtAngle();
-			} else {
-				projectile = getArrow(world, entity, stack, ammo);
-				if (p_40901_ || p_40904_ != 0.0F) {
-					((EntityArrow)projectile).pickupStatus = EntityArrow.PickupStatus.CREATIVE_ONLY;
-				}
+	private static void shootProjectile(World world, EntityLivingBase entity, ItemStack stack, ItemStack ammo, float p_40900_, boolean p_40901_, float p_40902_, float p_40903_, float p_40904_) {
+		if (world.isRemote) return;
+		boolean isFirework = ammo.getItem() == Items.FIREWORKS;
+		Entity projectile;
+		if (isFirework) {
+			projectile = new EntityFireworkRocket(world, entity.posX, entity.posY + entity.getEyeHeight() - 0.15F, entity.posZ, ammo);
+			((IFireworksProjectile)projectile).setOwner(entity);
+			((IFireworksProjectile)projectile).setShotAtAngle();
+		} else {
+			projectile = getArrow(world, entity, stack, ammo);
+			if (p_40901_ || p_40904_ != 0.0F) {
+				((EntityArrow)projectile).pickupStatus = EntityArrow.PickupStatus.CREATIVE_ONLY;
 			}
-
-			if (entity instanceof ICrossbowAttackMob) {
-				ICrossbowAttackMob crossbowattackmob = (ICrossbowAttackMob)entity;
-				crossbowattackmob.shootCrossbowProjectile(entity, crossbowattackmob.getTarget(), projectile, p_40904_, 1.6F);
-			} else {
-				Vec3d vec31 = entity.getLook(1.0F);
-				Vector3f angle = new Vector3f((float) vec31.x, (float) vec31.y, (float) vec31.z);
-				Vec3d vec3 = entity.getLook(1.0F);
-				Vector3f vector3f = new Vector3f((float) vec3.x, (float) vec3.y, (float) vec3.z);
-				vector3f.angle(angle);
-				((IProjectile)projectile).shoot(vector3f.x, vector3f.y, vector3f.z, p_40902_, p_40903_);
-			}
-
-			stack.damageItem(isFirework ? 3 : 1, entity);
-			world.spawnEntity(projectile);
-			world.playSound(null, entity.posX, entity.posY, entity.posZ, RaidsSoundEvents.CROSSBOW_SHOOT, SoundCategory.PLAYERS, 1.0F, p_40900_);
 		}
+		if (entity instanceof ICrossbowAttackMob) {
+			ICrossbowAttackMob crossbowattackmob = (ICrossbowAttackMob)entity;
+			crossbowattackmob.shootCrossbowProjectile(entity, crossbowattackmob.getTarget(), projectile, p_40904_, 1.6F);
+		} else {
+			Vec3d look = entity.getLook(1.0F);
+			float angle =  p_40904_ * ((float)Math.PI / 180F);
+			double s = (float) Math.sin(angle);
+			float c = (float) Math.cos(angle);
+			((IProjectile)projectile).shoot(look.x * c - look.z * s, look.y, look.x * s + look.z * c, p_40902_, p_40903_);
+		}
+		stack.damageItem(isFirework ? 3 : 1, entity);
+		world.spawnEntity(projectile);
+		world.playSound(null, entity.posX, entity.posY, entity.posZ, RaidsSoundEvents.CROSSBOW_SHOOT, SoundCategory.PLAYERS, 1.0F, p_40900_);
 	}
 
-	private static EntityArrow getArrow(World p_40915_, EntityLivingBase p_40916_, ItemStack p_40917_, ItemStack p_40918_) {
-		ItemArrow arrowitem = (ItemArrow) (p_40918_.getItem() instanceof ItemArrow ? p_40918_.getItem() : Items.ARROW);
-		EntityArrow abstractarrow = arrowitem.createArrow(p_40915_, p_40918_, p_40916_);
-		if (p_40916_ instanceof EntityPlayer) {
-			abstractarrow.setIsCritical(true);
-		}
-		ICrossbowArrow crossbowArrow = (ICrossbowArrow) abstractarrow;
+	private static EntityArrow getArrow(World world, EntityLivingBase entity, ItemStack crossbow, ItemStack ammo) {
+		ItemArrow arrowitem = (ItemArrow) (ammo.getItem() instanceof ItemArrow ? ammo.getItem() : Items.ARROW);
+		EntityArrow arrow = arrowitem.createArrow(world, ammo, entity);
+		if (entity instanceof EntityPlayer) arrow.setIsCritical(true);
+		ICrossbowArrow crossbowArrow = (ICrossbowArrow) arrow;
 		crossbowArrow.setShotFromCrossbow(true);
-		int i = EnchantmentHelper.getEnchantmentLevel(RaidsContent.PIERCING, p_40917_);
-		if (i > 0) {
-			crossbowArrow.setPierceLevel((byte)i);
-		}
-
-		return abstractarrow;
+		int i = EnchantmentHelper.getEnchantmentLevel(RaidsContent.PIERCING, crossbow);
+		if (i > 0) crossbowArrow.setPierceLevel((byte)i);
+		return arrow;
 	}
 
-	public static void performShooting(World world, EntityLivingBase entity, EnumHand hand, ItemStack stack, float p_40892_, float p_40893_) {
+	public static void performShooting(World world, EntityLivingBase entity, ItemStack stack, float p_40892_, float p_40893_) {
 		if (entity instanceof EntityPlayer && ForgeEventFactory.onArrowLoose(stack, world, (EntityPlayer) entity, 1, true) < 0) return;
 		List<ItemStack> list = getChargedProjectiles(stack);
 		float[] afloat = getShotPitches(entity.getRNG());
-
 		for(int i = 0; i < list.size(); ++i) {
 			ItemStack itemstack = list.get(i);
 			boolean flag = entity instanceof EntityPlayer && ((EntityPlayer)entity).capabilities.isCreativeMode;
 			if (!itemstack.isEmpty()) {
-				if (i == 0) {
-					shootProjectile(world, entity, hand, stack, itemstack, afloat[i], flag, p_40892_, p_40893_, 0.0F);
-				} else if (i == 1) {
-					shootProjectile(world, entity, hand, stack, itemstack, afloat[i], flag, p_40892_, p_40893_, -10.0F);
-				} else if (i == 2) {
-					shootProjectile(world, entity, hand, stack, itemstack, afloat[i], flag, p_40892_, p_40893_, 10.0F);
-				}
+				if (i == 0) shootProjectile(world, entity, stack, itemstack, afloat[i], flag, p_40892_, p_40893_, 0.0F);
+				else if (i == 1) shootProjectile(world, entity, stack, itemstack, afloat[i], flag, p_40892_, p_40893_, -10.0F);
+				else if (i == 2) shootProjectile(world, entity, stack, itemstack, afloat[i], flag, p_40892_, p_40893_, 10.0F);
 			}
 		}
 		onCrossbowShot(world, entity, stack);
@@ -274,8 +243,8 @@ public class ItemCrossbow extends Item {
 	public void onUsingTick(ItemStack stack, EntityLivingBase entity, int count) {
 	      if (!entity.world.isRemote) {
 	         int i = EnchantmentHelper.getEnchantmentLevel(RaidsContent.QUICK_CHARGE, stack);
-	         SoundEvent soundevent = getStartSound(0);
-	         SoundEvent soundevent1 = i == 0 ? RaidsSoundEvents.CROSSBOW_LOADING_MIDDLE : null;
+	         SoundEvent startSound = getStartSound(0);
+	         SoundEvent middleSound = i == 0 ? RaidsSoundEvents.CROSSBOW_LOADING_MIDDLE : null;
 	         float f = (float)(stack.getMaxItemUseDuration() - count) / (float)getChargeDuration(stack);
 	         if (f < 0.2F) {
 	            startSoundPlayed = false;
@@ -284,13 +253,13 @@ public class ItemCrossbow extends Item {
 
 	         if (f >= 0.2F && !startSoundPlayed) {
 	            startSoundPlayed = true;
-				 entity.world.playSound(null, entity.posX, entity.posY, entity.posZ, soundevent, entity instanceof EntityPlayer ?
+				 entity.world.playSound(null, entity.posX, entity.posY, entity.posZ, startSound, entity instanceof EntityPlayer ?
 						 SoundCategory.PLAYERS : SoundCategory.HOSTILE, 0.5F, 1.0F);
 	         }
 
-	         if (f >= 0.5F && soundevent1 != null && !this.midLoadSoundPlayed) {
+	         if (f >= 0.5F && middleSound != null && !this.midLoadSoundPlayed) {
 				 midLoadSoundPlayed = true;
-				 entity.world.playSound(null, entity.posX, entity.posY, entity.posZ, soundevent1, entity instanceof EntityPlayer ?
+				 entity.world.playSound(null, entity.posX, entity.posY, entity.posZ, middleSound, entity instanceof EntityPlayer ?
 						 SoundCategory.PLAYERS : SoundCategory.HOSTILE, 0.5F, 1.0F);
 	         }
 	      }
@@ -306,8 +275,8 @@ public class ItemCrossbow extends Item {
 		return i == 0 ? 25 : 25 - 5 * i;
 	}
 
-	private SoundEvent getStartSound(int p_40852_) {
-		switch(p_40852_) {
+	private SoundEvent getStartSound(int quick_charge_level) {
+		switch(quick_charge_level) {
 			case 1:
 				return RaidsSoundEvents.CROSSBOW_QUICK_CHARGE_1;
 			case 2:
